Refactor MongoDB S3 backup
